Chore : move to splash package
