+++ feature/intro/impl/src/main/java/com/teamyg/parfait/feature/intro/impl/splash/SplashRoute.kt
@@ -1,4 +1,4 @@
-package com.teamyg.parfait.feature.intro.impl
+package com.teamyg.parfait.feature.intro.impl.splash
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
